Week-2 (#26)
The BIC for the linear model (27.035052) is smaller than the BIC for the cubic model (27.416137), what suggests that the linear model is preferable over the cubic one. However, this is not the only parameter that should be evaluated: an F-test, the r² value and the residuals plot should be used in conjunction with BIC to decide which model explains better the relationship between the variables.